a lot of fixes and wscripts

+++ public/tier0/basetypes.h
@@ -405,14 +405,14 @@ inline T AlignValue( T val, uintp alignment )
 
 #ifdef __cplusplus
 
-inline unsigned long& FloatBits( vec_t& f )
+inline unsigned int& FloatBits( vec_t& f )
 {
-	return *reinterpret_cast<unsigned long*>(&f);
+	return *reinterpret_cast<unsigned int*>(&f);
 }
 
-inline unsigned long const& FloatBits( vec_t const& f )
+inline unsigned int const& FloatBits( vec_t const& f )
 {
-	return *reinterpret_cast<unsigned long const*>(&f);
+	return *reinterpret_cast<unsigned int const*>(&f);
 }
 
 inline vec_t BitsToFloat( unsigned long i )
